Use STI for Schedule model
