+++ app/models/serverspec_schedule.rb
@@ -6,9 +6,7 @@
 # http://opensource.org/licenses/mit-license.php
 #
 
-class ServerspecSchedule < ActiveRecord::Base
-  include Concerns::Schedule
-
+class ServerspecSchedule < Schedule
   belongs_to :resource, foreign_key: 'physical_id', primary_key: 'physical_id'
 
   @@job_class_name = PeriodicServerspecJob.to_s.freeze
