--- db/migrate/20150619084826_create_yum_schedules.rb
@@ -1,13 +1,0 @@
-class CreateYumSchedules < ActiveRecord::Migration
-  def change
-    create_table :yum_schedules do |t|
-      t.string  :physical_id, null: false
-      t.boolean :enabled,     null: false, default: false
-      t.integer :frequency,   null: true
-      t.integer :day_of_week, null: true
-      t.integer :time,        null: true
-
-      t.timestamps null: false
-    end
-  end
-end
