+++ db/migrate/20150702041425_rename_table_serverspec_schedules_to_schedules.rb
@@ -1,0 +1,5 @@
+class RenameTableServerspecSchedulesToSchedules < ActiveRecord::Migration
+  def change
+    rename_table :serverspec_schedules, :schedules
+  end
+end
