+++ db/migrate/20150702041553_add_type_to_schedules.rb
@@ -1,0 +1,16 @@
+class AddTypeToSchedules < ActiveRecord::Migration
+  def up
+    add_column :schedules, :type, :string
+
+    Schedule.all.each do |e|
+      e.type = 'ServerspecSchedule'
+      e.save!
+    end
+
+    change_column :schedules, :type, :string, null: false
+  end
+
+  def down
+    remove_column :schedules, :type
+  end
+end
